+++ default.nix
@@ -7,6 +7,7 @@
 , useFastWeak ? true
 , useReflexOptimizer ? false
 , useTextJSString ? true # Use an implementation of "Data.Text" that uses the more performant "Data.JSString" from ghcjs-base under the hood.
+, useWebkit2Gtk ? false # Enable webkit2gtk to build reflex-dom desktop apps
 , __useTemplateHaskell ? true # Deprecated, just here until we remove feature from reflex and stop CIing it
 , __useNewerCompiler ? true
 , iosSdkVersion ? "16.1"
@@ -19,6 +20,7 @@
 
 let iosSupport = system == "x86_64-darwin";
     androidSupport = lib.elem system [ "x86_64-linux" ];
+    ghc86Support = lib.elem system ["x86_64-linux" "x86_64-darwin"];
 
     xcodeVer = {
       "16.1" = "14.1";
@@ -49,13 +51,19 @@ let iosSupport = system == "x86_64-darwin";
             bootPkgs = super.haskell.packages.ghc865Binary // {
               happy = super.haskell.packages.ghc865Binary.happy_1_19_12;
             };
+            useLdGold = !(self.stdenv.targetPlatform.isAarch32) && self.stdenv.hostPlatform.useAndroidPrebuilt;
+            enableDocs = false;
+            enableHaddockProgram = false;
           };
           ghcSplices-8_10 = (super.haskell.compiler.ghc8107.override {
             # New option for GHC 8.10. Explicitly enable profiling builds
             enableProfiledLibs = true;
-            bootPkgs = super.haskell.packages.ghc865Binary // {
-              happy = super.haskell.packages.ghc865Binary.happy_1_19_12;
-            };
+            #enableShared = self.stdenv.hostPlatform == self.stdenv.targetPlatform;
+            #enableShared = false;
+            bootPkgs = if (super.stdenv.hostPlatform.isAarch64) then (super.haskell.packages.ghc8107Binary // {
+              happy = super.haskell.packages.ghc8107Binary.happy_1_19_12;
+            }) else
+            (super.haskell.packages.ghc865Binary // { happy = super.haskell.packages.ghc865Binary.happy_1_19_12; });
           }).overrideAttrs (drv: {
             src = nixpkgs.hackGet ./haskell-overlays/splices-load-save/dep/ghc-8.10;
             # When building from the ghc git repo, ./boot must be run before configuring, whereas
@@ -94,7 +102,8 @@ let iosSupport = system == "x86_64-darwin";
             useFastWeak useReflexOptimizer enableLibraryProfiling enableTraceReflexEvents
             useTextJSString enableExposeAllUnfoldings __useTemplateHaskell
             haskellOverlaysPre
-            haskellOverlaysPost;
+            haskellOverlaysPost
+            useWebkit2Gtk;
           inherit ghcSavedSplices-8_6 ghcSavedSplices-8_10;
         };
       };
@@ -105,13 +114,13 @@ let iosSupport = system == "x86_64-darwin";
         libiconv = super.darwin.libiconv.overrideAttrs (_:
           lib.optionalAttrs (self.stdenv.hostPlatform != self.stdenv.buildPlatform) {
             postInstall = "rm $out/include/libcharset.h $out/include/localcharset.h";
-            configureFlags = ["--disable-shared" "--enable-static"];
+            configureFlags = ["--enable-shared" "--enable-static"];
           });
-      };
+        };
       zlib = super.zlib.override (lib.optionalAttrs
         (self.stdenv.hostPlatform != self.stdenv.buildPlatform)
-        { static = true; shared = false; });
-    };
+        { static = true; shared = true; });
+      };
 
     mobileGhcOverlay = import ./nixpkgs-overlays/mobile-ghc { inherit lib; };
 
@@ -120,6 +129,7 @@ let iosSupport = system == "x86_64-darwin";
     nixpkgsArgs = {
       inherit system;
       overlays = [
+        (import ./nixpkgs-overlays/ghc.nix { inherit lib; })
         hackGetOverlay
         bindHaskellOverlays
         forceStaticLibs
@@ -127,26 +137,45 @@ let iosSupport = system == "x86_64-darwin";
         mobileGhcOverlay
         allCabalHashesOverlay
         (self: super: {
-          binutils-unwrapped = super.binutils-unwrapped.override {
-            autoreconfHook = lib.optional self.stdenv.buildPlatform.isDarwin super.autoreconfHook269;
+
+          runtimeShellPackage = if (self.stdenv.hostPlatform.isGhcjs || self.stdenv.targetPlatform.isiOS)
+            then super.buildPackages.runtimeShellPackage
+            else super.runtimeShellPackage;
+
+          polkit = super.polkit.override {
+            gobject-introspection = super.gobject-introspection-unwrapped;
           };
 
-          # Bump ios-deploy
-          # - for faster deployments
-          # - fixes debug deploy with iOS 16/macos 12.3/ xcode 13.4.1
-          darwin = super.darwin // {
-            ios-deploy = super.darwin.ios-deploy.overrideAttrs (_: {
-              version = "HEAD";
+          darwin = super.darwin.overrideScope (dself: dsuper: {
+            ios-deploy = dsuper.ios-deploy.overrideAttrs (_: {
+              version = "1.12.2";
               src = self.fetchFromGitHub {
                 owner = "ios-control";
                 repo = "ios-deploy";
-                rev = "b3254438719b6bc82ceab1f630e7d642a9acfac5"; # unreleased
-                sha256 = "W45Qjr3xqvDWieLBgt4//nthxxcc3hgrJNrpSk7vWj8=";
+                rev = "ed7de7792d28a5110242748649047a95c95ea917";
+                sha256 = "082w7j490khfpbv1diwrjixjbg9g93wdr2khyzdhv8xmzzwq4lad";
               };
             });
+          });
+          openjdk16-bootstrap = super.openjdk16-bootstrap.override {
+            gtkSupport = false;
           };
+          adoptopenjdk-hotspot-bin-16 = super.adoptopenjdk-hotspot-bin-16.override {
+            gtkSupport = false;
+          };
+
+          sqlite = super.sqlite.overrideAttrs (old: lib.optionalAttrs (self.stdenv.hostPlatform.useAndroidPrebuilt or false) {
+            postBuild = ''
+              mkdir -p $debug
+            '';
+          });
+
+          libiconv = super.libiconv.overrideAttrs (old: lib.optionalAttrs (self.stdenv.hostPlatform.useAndroidPrebuilt or false) {
+            configureFlags = [ "--disable-shared" "--enable-static" ];
+          });
+
+          libffi = if (self.stdenv.hostPlatform.useAndroidPrebuilt or false) then super.libffi_3_3 else super.libffi;
         })
-        (import ./nixpkgs-overlays/ghc.nix { inherit lib; })
       ] ++ nixpkgsOverlays;
       config = config // {
         permittedInsecurePackages = (config.permittedInsecurePackages or []) ++ [
@@ -166,24 +195,41 @@ let iosSupport = system == "x86_64-darwin";
     wasmCross = nixpkgs.hackGet ./wasm-cross;
     webGhcSrc = (import (wasmCross + /webghc.nix) { inherit fetchgit; }).ghc8107SplicesSrc;
     nixpkgsCross = {
+      # NOTE(Dylan Green):
+      # sdkVer 30 is the minimum for android, else we have to use libffi 3.3
+      # bionic doesn't support/expose memfd_create before sdk30
+      # https://android.googlesource.com/platform/bionic/+/refs/heads/master/docs/status.md
+      # Look for "new libc functions in R (API Level 30):", memfd_create will be one of the functions /
+      # symbols we need to build newer libffi
+      # This means we'll drop all SDKs pre-30
+
+      # NOTE(Dylan Green):
+      # We don't want to use "isStatic" here as we still rely on shared-objects
+      # adding "isStatic" completely disables generating most SOs, and we still need them
+      # for libffi (at the very least). Currently the big issues are caused by the linker attempting (and failing)
+      # to link against a dynamic crtbegin.o (crtbegin.c) bionic does provide a static crtbegin, although the linker
+      # defaults to a dynamic version
+
+      # TODO(Dylan Green):
+      # Look into making this a proper static build up into "reflex-todomvc"
       android = lib.mapAttrs (_: args: nixpkgsFunc (nixpkgsArgs // args)) rec {
         aarch64 = {
-          crossSystem = lib.systems.examples.aarch64-android-prebuilt //
-          { isStatic = true; };
-          sdkVer = "30";
+          crossSystem = lib.systems.examples.aarch64-android-prebuilt // {
+            #isStatic = true;
+            sdkVer = "30";
+          };
         };
         aarch32 = {
           crossSystem = lib.systems.examples.armv7a-android-prebuilt // {
-            isStatic = true;
-            # Choose an old version so it's easier to find phones to test on
-            sdkVer = "23";
+            #isStatic = true;
+            sdkVer = "30";
           };
         };
       };
       ios = lib.mapAttrs (_: args: nixpkgsFunc (nixpkgsArgs // args)) rec {
         simulator64 = {
           crossSystem = lib.systems.examples.iphone64-simulator // {
-            #isStatic = true;
+            isStatic = true;
             sdkVer = iosSdkVersion;
             inherit xcodeVer;
           };
@@ -229,12 +275,6 @@ let iosSupport = system == "x86_64-darwin";
       override = new: makeRecursivelyOverridable (x.override (old: (combineOverrides old new)));
     };
 
-    cabal2nixResult = src: builtins.trace "cabal2nixResult is deprecated; use ghc.haskellSrc2nix or ghc.callCabal2nix instead" (ghc.haskellSrc2nix {
-      name = "for-unknown-package";
-      src = "file://${src}";
-      sha256 = null;
-    });
-
   ghcSavedSplices = if __useNewerCompiler then ghcSavedSplices-8_10 else ghcSavedSplices-8_6;
   ghcSavedSplices-8_6 = (makeRecursivelyOverridable nixpkgs.haskell.packages.integer-simple.ghcSplices-8_6).override {
     overrides = lib.foldr lib.composeExtensions (_: _: {}) (let
@@ -269,8 +309,7 @@ let iosSupport = system == "x86_64-darwin";
   ghcjs = if __useNewerCompiler then ghcjs8_10 else ghcjs8_6;
   ghcjs8_6 = (makeRecursivelyOverridable (nixpkgsCross.ghcjs.haskell.packages.ghcjs86.override (old: {
     ghc = old.ghc.override {
-      bootPkgs = with nixpkgsCross.ghcjs.buildPackages.haskell.packages;
-        ghc865 // { happy = ghc865.callHackage "happy" "1.19.9" {}; };
+      bootPkgs = old.ghc.bootPkgs // { happy = old.ghc.bootPkgs.happy_1_19_12; };
       cabal-install = import ./haskell-overlays/ghcjs-8.6/cabal-install.nix { inherit nixpkgs; };
       ghcjsSrc = import ./haskell-overlays/ghcjs-8.6/src.nix {
         inherit (nixpkgs.stdenvNoCC) mkDerivation;
@@ -514,7 +553,6 @@ in let this = rec {
       cabal-install
       ghcid
       hasktags
-      hlint
       stylish-haskell # Recent stylish-haskell only builds with AMP in place
       reflex-ghci
       ;
@@ -589,7 +627,7 @@ in let this = rec {
         iosReflexTodomvc iosSimulatorReflexTodomvc
       ];
 
-  inherit cabal2nixResult system androidSupport iosSupport;
+  inherit system androidSupport iosSupport ghc86Support;
   project = args: import ./project this (args ({ pkgs = nixpkgs; } // this));
   tryReflexShell = pinBuildInputs ("shell-" + system) tryReflexPackages;
   ghcjsExternsJs = ./ghcjs.externs.js;