+++ haskell-overlays/default.nix
@@ -6,6 +6,7 @@
 , enableLibraryProfiling
 , enableTraceReflexEvents
 , useTextJSString
+, useWebkit2Gtk
 , enableExposeAllUnfoldings
 , __useTemplateHaskell
 , ghcSavedSplices-8_6
@@ -50,6 +51,10 @@ rec {
 
     (optionalExtension enableExposeAllUnfoldings exposeAllUnfoldings)
 
+    #(NEW;Dylan Green):
+    # We no longer need to set gold as "lld" is default on the
+    # android toolchain now
+    #(OLD;Dylan Green):
     # Force "gold" on Android due to a linker bug on bfd
     # Also force -fPIC on for Android, we need it either way
 
@@ -58,22 +63,27 @@ rec {
     # arm* needs the same linker options, x86* -> arm* does not
 
     (optionalExtension (super.ghc.stdenv.targetPlatform.isAndroid or false) (self: super:
-      {
+    {
         mkDerivation = drv: super.mkDerivation (drv // {
           buildFlags = [
-            "--ghc-option=-optl-fuse-ld=gold"
-            "--ld-option=-fuse-ld=gold"
             "--ld-option=-fPIE"
             "--ld-option=-pie"
             "--ghc-option=-fPIC"
             "--ghc-option=-fPIE"
           ] ++ (drv.buildFlags or [ ]);
 
-          configureFlags = [
-            "--with-ld=${super.ghc.stdenv.targetPlatform.config}-ld.gold"
-          ] ++ (drv.configureFlags or [ ]);
+          configureFlags = [ ] ++ (drv.configureFlags or [ ]);
         });
       }))
+
+    # TODO(Dylan): Add this casing to the compiler patch
+    (optionalExtension (super.ghc.stdenv.targetPlatform.isiOS && (super.ghc.stdenv.targetPlatform.isx86_64 || super.ghc.version == "8.6.5")) (self: super: {
+      mkDerivation = drv: super.mkDerivation (drv // {
+        buildFlags = (drv.buildFlags or []) ++ [
+          "--ghc-option=-fwhole-archive-hs-libs"
+        ];
+      });
+    }))
 
     combined-any
     (optionalExtension (!(super.ghc.isGhcjs or false)) combined-ghc)
@@ -139,6 +149,17 @@ rec {
     (optionalExtension useTextJSString textJSString-8_10)
     (optionalExtension useTextJSString ghcjs-textJSString-8_10)
     (optionalExtension useFastWeak ghcjs-fast-weak_8_10)
+    (self: super: rec {
+      mkDerivation = drv: super.mkDerivation (drv // {
+        setupHaskellDepends = (drv.setupHaskellDepends or []) ++ [
+          nixpkgs.buildPackages.stdenv.cc
+        ];
+        # This is ugly
+        preConfigure = (drv.preConfigure or "") + ''
+          export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:${nixpkgs.buildPackages.gmp}/lib:${nixpkgs.buildPackages.libffi}/lib
+        '';
+      });
+    })
   ]
     self
     super;
@@ -151,6 +172,7 @@ rec {
     inherit
       haskellLib lib nixpkgs thunkSet fetchFromGitHub fetchFromBitbucket hackGet
       useFastWeak useReflexOptimizer enableTraceReflexEvents enableLibraryProfiling __useTemplateHaskell
+      useWebkit2Gtk
       ;
   };
   exposeAllUnfoldings = import ./expose-all-unfoldings.nix { };