+++ haskell-overlays/ghcjs-8.6/cabal-install.nix
@@ -4,6 +4,7 @@
   (self: super: {
     Cabal = self.Cabal_2_4_1_0;
     base16-bytestring = self.base16-bytestring_0_1_1_7;
+    random = self.callHackage "random" "1.1" {};
     hackage-security = nixpkgs.haskell.lib.dontCheck
       (nixpkgs.haskell.lib.doJailbreak
         (self.callHackage "hackage-security" "0.5.3.0" {}));