+++ haskell-overlays/profiling.nix
@@ -4,10 +4,14 @@
 
 with haskellLib;
 
+let
+  preventMobileProfiling = self: (!self.ghc.stdenv.targetPlatform.isiOS) && enableLibraryProfiling;
+in
+
 self: super: {
 
   # Override mkDerivation to inherit global settings
   mkDerivation = expr: super.mkDerivation (expr // {
-    inherit enableLibraryProfiling;
+    enableLibraryProfiling = preventMobileProfiling self;
   });
 }