+++ haskell-overlays/reflex-packages/default.nix
@@ -2,6 +2,7 @@
 , lib, nixpkgs
 , thunkSet, fetchFromGitHub, fetchFromBitbucket, hackGet
 , useFastWeak, useReflexOptimizer, enableTraceReflexEvents, enableLibraryProfiling, __useTemplateHaskell
+, useWebkit2Gtk
 }:
 
 with haskellLib;
@@ -17,6 +18,9 @@ let
 
   reflexOptimizerFlag = lib.optional (useReflexOptimizer && (self.ghc.cross or null) == null) "-fuse-reflex-optimizer";
   useTemplateHaskellFlag = lib.optional (!__useTemplateHaskell) "-f-use-template-haskell";
+  useWebkit2GtkFlag = if useWebkit2Gtk
+    then ["-fwebkit2gtk"]
+    else ["-f-webkit2gtk"] ++ lib.optional ((nixpkgs.stdenv.hostPlatform.isLinux or false) && !nixpkgs.stdenv.hostPlatform.useAndroidPrebuilt) "-fuse-warp"; # Enable warp on linux if webkit2gtk is disabled. Other platforms have other default runners
 
   inherit (nixpkgs) stdenv;
   # Older chromium for reflex-dom-core test suite
@@ -46,7 +50,20 @@ in
     (lib.optional useFastWeak "-ffast-weak")
   ])) {};
 
-  reflex-todomvc = self.callPackage self._dep.reflex-todomvc {};
+  reflex-todomvc =
+    let
+      flags =
+        if useWebkit2Gtk && nixpkgs.stdenv.hostPlatform.isLinux
+        then [ "-f-warp" "-f-webkitgtk" "-f-wkwebview" ]
+        else if (nixpkgs.stdenv.hostPlatform.isLinux && !nixpkgs.stdenv.hostPlatform.useAndroidPrebuilt)
+        then [ "-fwarp" "-f-webkitgtk" "-f-wkwebview" "-f-webkit2gtk" ]
+        else if self.ghc.stdenv.targetPlatform.isiOS
+        then [ "-f-webkit2gtk" "-f-warp" "-f-webkitgtk" ]
+        else if nixpkgs.stdenv.hostPlatform.isDarwin
+        then [ "-fwkwebview" "-f-webkit2gtk" "-f-webkitgtk" ]
+        else [];
+    in
+      (haskellLib.doJailbreak (self.callCabal2nixWithOptions "reflex-todomvc" self._dep.reflex-todomvc (lib.concatStringsSep " " flags) {}));
   reflex-aeson-orphans = self.callCabal2nix "reflex-aeson-orphans" self._dep.reflex-aeson-orphans {};
 
   # The tests for reflex-dom-core are not deterministic, disable them, and run them manually
@@ -54,6 +71,7 @@ in
     inherit (self) ghc;
     noGcTest = stdenv.hostPlatform.system != "x86_64-linux"
             || stdenv.hostPlatform != stdenv.buildPlatform
+            || stdenv.targetPlatform.isiOS
             || (ghc.isGhcjs or false);
   in haskellLib.overrideCabal
     (self.callCabal2nixWithOptions "reflex-dom-core" (reflexDomRepo + "/reflex-dom-core") (lib.concatStringsSep " " (lib.concatLists [
@@ -96,11 +114,23 @@ in
       '' + (drv.preCheck or "");
     });
 
-  reflex-dom =
-    self.callCabal2nixWithOptions "reflex-dom" (reflexDomRepo + "/reflex-dom") (lib.concatStringsSep " " (lib.concatLists [
+    reflex-dom = haskellLib.doJailbreak (haskellLib.overrideCabal (self.callCabal2nixWithOptions "reflex-dom" (reflexDomRepo + "/reflex-dom") (lib.concatStringsSep " " (lib.concatLists [
       reflexOptimizerFlag
       useTemplateHaskellFlag
-    ])) {};
+      useWebkit2GtkFlag
+    ])) { }) (drv: {
+      preConfigure = (drv.preConfigure or "") + ''
+        sed -i 's|aeson >=1.4 && <1.6|aeson -any|g' *.cabal
+      '';
+
+      libraryHaskellDepends = [
+        self.reflex
+        self.reflex-dom-core
+        self.aeson
+      ] ++ lib.optional (nixpkgs.stdenv.hostPlatform.useAndroidPrebuilt or false) self.android-activity
+        ++ lib.optional (nixpkgs.stdenv.hostPlatform.isDarwin or false) self.jsaddle-wkwebview
+        ++ lib.optional ((nixpkgs.stdenv.hostPlatform.isLinux or false) && !nixpkgs.stdenv.hostPlatform.useAndroidPrebuilt) self.jsaddle-warp;
+    }));
 
   chrome-test-utils = self.callCabal2nix "chrome-test-utils" (reflexDomRepo + "/chrome-test-utils") {};
 
@@ -108,7 +138,7 @@ in
   ## Terminal / Conventional OS
   ##
 
-  reflex-vty = self.callCabal2nix "reflex-vty" self._dep.reflex-vty {};
+  reflex-vty = haskellLib.doJailbreak (self.callCabal2nix "reflex-vty" self._dep.reflex-vty {});
   reflex-process = self.callCabal2nix "reflex-process" self._dep.reflex-process {};
   reflex-fsnotify = self.callCabal2nix "reflex-fsnotify" self._dep.reflex-fsnotify {};
 
@@ -144,7 +174,7 @@ in
     preConfigure = "substituteInPlace jsaddle-warp.cabal --replace 'aeson >=0.8.0.2 && <2.1' aeson";
   });
 
-  jsaddle-dom = self.callCabal2nix "jsaddle-dom" self._dep.jsaddle-dom {};
+  jsaddle-dom = doJailbreak (self.callCabal2nix "jsaddle-dom" self._dep.jsaddle-dom {});
   jsaddle-wasm = self.callCabal2nix "jsaddle-wasm" (hackGet (wasmCross + "/jsaddle-wasm")) {};
   ghcjs-dom = self.callCabal2nix "ghcjs-dom" (self._dep.ghcjs-dom + "/ghcjs-dom") {};
   ghcjs-dom-jsaddle = self.callCabal2nix "ghcjs-dom-jsaddle" (self._dep.ghcjs-dom + "/ghcjs-dom-jsaddle") {};
@@ -178,7 +208,7 @@ in
   commutative-semigroups = self.callCabal2nix "commutative-semigroups" self._dep.commutative-semigroups {};
   witherable = self.callHackage "witherable" "0.4.2" {};
 
-  webdriver = self.callHackage "webdriver" "0.9.0.1" {};
+  webdriver = markUnbroken (self.callHackage "webdriver" "0.9.0.1" {});
 
   # Not on Hackage yet
   # Version 1.2.1 not on Hackage yet